# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: FlowForge Control Plane API
  version: 0.13.0
  description: |
    Versioned control-plane API. Conventions:

    - Correlation: every response sets `X-Request-ID`. A caller value is
      accepted only when it is 16–128 ASCII letters, digits, or hyphens;
      otherwise the API generates one. The same id is returned in problem
      documents as `request_id` and in structured request logs.
    - Errors: 4xx/5xx responses use RFC 9457 `application/problem+json`
      (`type`, `title`, `status`, `detail`, `instance`) plus `code` and
      `request_id`. Problem details never echo request bodies, credentials,
      or other secret material.
      - Invalid requests use `invalid-request` (400) or `request-too-large` (413).
      Invalid workflow YAML uses `invalid-workflow` (400) with an `errors`
      array (path, line, column, code, message). Authenticated-but-failed
      flows use `unauthenticated` (401) and `forbidden` (403). Unique
      identity collisions, draft revision mismatches, and duplicate
      published digests use `conflict` (409).
    - Browser clients authenticate with `ff_session` / `ff_csrf` cookies from
      `POST /session`. State-changing cookie requests require `X-CSRF-Token`.
      Non-browser callers may keep using `X-FlowForge-Issuer` /
      `X-FlowForge-Subject`. A session cookie wins; conflicting identity
      headers fail closed.
    - Workspace identity is server-derived from `(tenant_id|tenant_slug,
...
servers:
...
paths:
...
  /executions/{executionId}/cancel:
    post:
      operationId: cancelExecution
      summary: Cancel an execution
      description: |
        Requires `execution.cancel` (operator/admin). Idempotent: a second
        cancel of an already-canceled run returns `200`. Terminal
        succeeded/failed/indeterminate runs are `409`. Viewers receive `403`.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
        - name: executionId
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmptyDispatchRequest"
      responses:
        "200":
          description: Canceled execution detail
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowExecution"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /executions/{executionId}/emergency-stop:
    post:
      operationId: emergencyStopExecution
      summary: Emergency-stop a running script execution
      description: |
        Requires `script.emergencyStop` (operator/admin) and an allowing
        kind=script policy when one is bound (`allowEmergencyStop` defaults
        to allow when omitted). Stops the runner. A queued/claimed job that
        has not heartbeated is `canceled`. A running job or `uncertain=true`
        is `indeterminate` until verified — never claimed succeeded or failed.
        Viewers receive `403`. Host-supplied workspace IDs are `400`.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
        - name: executionId
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptEmergencyStopRequest"
      responses:
        "200":
          description: Stopped execution detail
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowExecution"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /executions/{executionId}/retry:
    post:
      operationId: retryExecution
      summary: Retry the latest failed or canceled step
      description: |
        Requires `workflow.execute`. Creates a new attempt + queued job.
        Core `data.*` / `flow.*` nodes are retry-safe (max 3 retries).
        Provider nodes and `indeterminate` outcomes are `409`.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
        - name: executionId
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetryRequest"
      responses:
        "201":
          description: New queued attempt
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetryResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /jobs/claim:
    post:
      operationId: claimJob
      summary: Claim the next queued job
      description: |
        Requires `workflow.execute`. Recovers expired leases to
        `indeterminate` first, then claims one queued job with
        `FOR UPDATE SKIP LOCKED`. Returns an HMAC `jobToken` bound to
        workspace, workflow version/digest, policy digest, fencing token,
        and expiry. `204` when the queue is empty.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimJobRequest"
      responses:
        "200":
          description: Claimed job and authenticated ticket
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimJobResponse"
        "204":
          description: No eligible job
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /jobs/recover:
    post:
      operationId: recoverJobs
      summary: Mark expired leases indeterminate
      description: Requires `workflow.execute`. Unsafe lease loss never silently retries a provider call.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
      responses:
        "200":
          description: Recovery count
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoverJobsResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /jobs/{jobId}/heartbeat:
    post:
      operationId: heartbeatJob
      summary: Extend a job lease
      description: Requires `workflow.execute` plus a valid `jobToken`. First heartbeat marks the job `running` (provider fence). Stale fencing tokens are `409`.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
        - $ref: "#/components/parameters/JobID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobActionRequest"
      responses:
        "200":
          description: Updated job
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchJobResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /jobs/{jobId}/release:
    post:
      operationId: releaseJob
      summary: Release a claimed job
      description: Requires `workflow.execute` plus a valid `jobToken`. Release before heartbeat requeues; release after heartbeat is `indeterminate`.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
        - $ref: "#/components/parameters/JobID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobActionRequest"
      responses:
        "200":
          description: Updated job
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchJobResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /jobs/{jobId}/complete:
    post:
      operationId: completeJob
      summary: Complete a claimed job
      description: Requires `workflow.execute` plus a matching `jobToken` and fencing token. Output is redacted before persist.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
        - $ref: "#/components/parameters/JobID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobActionRequest"
      responses:
        "200":
          description: Completed job
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchJobResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /jobs/{jobId}/fail:
    post:
      operationId: failJob
      summary: Fail a claimed job
      description: Requires `workflow.execute` plus a matching `jobToken` and fencing token. Error is redacted before persist.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
        - $ref: "#/components/parameters/JobID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobActionRequest"
      responses:
        "200":
          description: Failed job
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchJobResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
...
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Correlation identifier. Accepted only when 16–128 ASCII letters, digits, or hyphens.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: ^[A-Za-z0-9-]{16,128}$
    Issuer:
      name: X-FlowForge-Issuer
      in: header
      required: false
      description: |
        Host identity issuer for non-browser callers. Browser clients should
        use `ff_session` instead. Ignored when a session cookie is present
        unless it conflicts (conflict fails closed). Does not authorize a workspace.
      schema:
        type: string
    Subject:
      name: X-FlowForge-Subject
      in: header
      required: false
      description: Host identity subject. Unique with issuer. Does not authorize a workspace.
      schema:
        type: string
...
    TenantID:
      name: X-FlowForge-Tenant-ID
      in: header
      required: false
      description: Tenant UUID used with workbench_key to resolve workspace identity.
      schema:
        type: string
        format: uuid
    TenantSlug:
      name: X-FlowForge-Tenant-Slug
      in: header
      required: false
      description: Tenant slug used with workbench_key when tenant id is omitted.
      schema:
        type: string
    WorkbenchKey:
      name: X-FlowForge-Workbench-Key
      in: header
      required: false
      description: Host workbench key. Combined with tenant identity, this is the only accepted workspace lookup.
      schema:
        type: string
    HostWorkspaceID:
      name: X-FlowForge-Workspace-ID
      in: header
      required: false
      description: Untrusted host-supplied workspace UUID. Rejected unless it matches the server-derived `(tenant, workbench_key)` workspace.
      schema:
        type: string
        format: uuid
...
  headers:
    XRequestID:
      description: Correlation identifier. Caller values are accepted only when 16–128 ASCII letters, digits, or hyphens.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: ^[A-Za-z0-9-]{16,128}$
  schemas:
...
    EmptyDispatchRequest:
      type: object
      properties:
        id: {type: string, description: "Rejected if present"}
        workspaceId: {type: string, description: "Rejected if present"}
...